+++ src/script.jl
@@ -1,37 +1,32 @@
 """
     Script(data) - object printed as text/javascript
 """
-struct Script{T}
-    content::T
+struct Script
+    content
 end
 
 Base.print(ep::EscapeProxy, x::Script) =
     print_script(ep.io, x.content)
 
 """
-    script(value)
-
-Convert a Julia value into something suitable for use within a
-`<script>` tag. By default, this will attempt to render the object
-as `"text/javascript"`. See also `print_script`.
-"""
-script(value) = Script(value)
-script(value::AbstractSet) = Script(collect(value))
-
-"""
     print_script(io, value)
 
-Render `value` as `"text/javascript"` to the given `io`.
+Show `value` as `"text/javascript"` to the given `io`, this provides
+some baseline functionality for built-in data types.
 
     - `nothing` becomes `undefined`
     - `missing` becomes `null`
     - `Symbol` becomes an unquoted name; no escaping
-    - `AbstractString` becomes a double-quoted string
     - `Bool` values are printed directly, as `true` or `false`
+    - `AbstractString` becomes a double-quoted string
+    - `AbstractVector` and `Tuple` become an array
+    - `Dict` and `NamedTuple` become a Javascript object.
 
 Numbers are simply printed, with a special case for Javascript's
 `Infinity` object; note that `NaN` is handled transparently.
 """
+print_script(io::IO, value) =
+    show(io, MIME"text/javascript"(), value)
 print_script(io::IO, ::Nothing) =
     print(io, "undefined")
 print_script(io::IO, ::Missing) =
@@ -82,10 +77,8 @@ function print_script(io::IO, value::AbstractString)
     final = lastindex(value)
     i = last = 1
     function emit(s::String)
-        if i > last
-            print(io, SubString(value, last, i - 1))
-            last = nextind(value, i)
-        end
+        print(io, SubString(value, last, i - 1))
+        last = nextind(value, i)
         print(io, s)
     end
     print(io, "\"")